MapViewをMapEditに改称、MapCanvasを別ディレクトリに独立
